ci(eslint): add import/no-cycle to prevent memory leak
