+++ .github/actions/pnpm-setup-node/action.yml
@@ -5,7 +5,7 @@ runs:
   steps:
     - uses: pnpm/action-setup@v2
       with:
-        version: 8.12.0
+        version: 8.12.1
     - uses: actions/setup-node@v4
       with:
         cache: "pnpm"
