+++ .vscode/settings.json
@@ -3,6 +3,5 @@
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": "explicit"
   },
-  "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"],
-  "typescript.tsdk": "node_modules/typescript/lib"
+  "eslint.validate": ["javascript", "javascriptreact", "typescript", "typescriptreact"]
 }
