+++ configs/eslint-config-ts/index.js
@@ -17,6 +17,7 @@ module.exports = {
   rules: {
     'sort-imports': ['error', { ignoreDeclarationSort: true }],
     'import/no-duplicates': 'error',
+    'import/no-cycle': 'error',
     'import/order': [
       'error',
       {
