+++ package.json
@@ -13,7 +13,7 @@
     "url": "https://github.com/suspensive/react.git"
   },
   "license": "MIT",
-  "packageManager": "pnpm@8.12.0",
+  "packageManager": "pnpm@8.12.1",
   "type": "module",
   "workspaces": [
     "configs/*",
