+++ packages/react-await/src/Await.tsx
@@ -1,7 +1,7 @@
 import type { FunctionComponent } from 'react'
 import { useMemo } from 'react'
 import { useSyncExternalStore } from 'use-sync-external-store/shim'
-import type { Tuple } from './types'
+import type { Tuple } from './utility-types'
 import { hashKey } from './utils'
 
 export type Key = Tuple
