+++ packages/react/src/hooks/index.ts
@@ -1,5 +1,4 @@
 export { useIsChanged } from './useIsChanged'
 export { useIsClient } from './useIsClient'
 export { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect'
-export { usePrevious } from './usePrevious'
 export { useTimeout } from './useTimeout'
