+++ packages/react/src/hooks/useIsChanged.ts
@@ -1,3 +1,13 @@
-import { usePrevious } from '.'
+import { useEffect, useRef } from 'react'
+
+export const usePrevious = <TValue>(value: TValue): TValue => {
+  const ref = useRef<TValue>(value)
+
+  useEffect(() => {
+    ref.current = value
+  }, [value])
+
+  return ref.current
+}
 
 export const useIsChanged = (value: unknown) => usePrevious(value) !== value
