--- packages/react/src/hooks/usePrevious.ts
@@ -1,11 +1,0 @@
-import { useEffect, useRef } from 'react'
-
-export const usePrevious = <TValue>(value: TValue): TValue => {
-  const ref = useRef<TValue>(value)
-
-  useEffect(() => {
-    ref.current = value
-  }, [value])
-
-  return ref.current
-}
