+++ packages/react/src/index.ts
@@ -11,9 +11,6 @@ export type { ErrorBoundaryGroupProps } from './ErrorBoundaryGroup'
 export type { DelayProps } from './Delay'
 
 // deprecated
-export { withSuspense } from './Suspense'
-export { withErrorBoundary } from './ErrorBoundary'
-export { withErrorBoundaryGroup } from './ErrorBoundaryGroup'
-export { withDelay } from './Delay'
+export { withSuspense, withDelay, withErrorBoundary, withErrorBoundaryGroup } from './wrap'
 export { AsyncBoundary, withAsyncBoundary } from './AsyncBoundary'
 export type { AsyncBoundaryProps } from './AsyncBoundary'
