+++ packages/react/src/utility-types/PropsWithoutChildren.ts
